revai/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
asyncpg
redis
python-dotenv
passlib
bcrypt==3.2.2
python-jose[cryptography]
pydantic[email]
httpx
